+++ apc/auth/auth.go
@@ -14,7 +14,7 @@ import (
 	"github.com/astronomer/astro-cli/houston"
 	"github.com/astronomer/astro-cli/pkg/input"
 	"github.com/astronomer/astro-cli/pkg/logger"
-	"github.com/astronomer/astro-cli/software/workspace"
+	"github.com/astronomer/astro-cli/apc/workspace"
 )
 
 const (
