+++ apc/deploy/testfiles/Dockerfile.privateImageRepo
