+++ cmd/apc/deploy_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/astronomer/astro-cli/houston"
 	testUtil "github.com/astronomer/astro-cli/pkg/testing"
-	"github.com/astronomer/astro-cli/software/deploy"
+	"github.com/astronomer/astro-cli/apc/deploy"
 )
 
 func execDeployCmd(args ...string) error {
