+++ cmd/apc/deployment_logs.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/astronomer/astro-cli/software/deployment"
+	"github.com/astronomer/astro-cli/apc/deployment"
 )
 
 const (
