+++ cmd/apc/deployment_teams.go
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/astronomer/astro-cli/houston"
-	"github.com/astronomer/astro-cli/software/deployment"
+	"github.com/astronomer/astro-cli/apc/deployment"
 )
 
 var deploymentRole string
