+++ cmd/apc/validation.go
@@ -9,7 +9,7 @@ import (
 	giturls "github.com/whilp/git-urls"
 
 	"github.com/astronomer/astro-cli/houston"
-	"github.com/astronomer/astro-cli/software/workspace"
+	"github.com/astronomer/astro-cli/apc/workspace"
 )
 
 var (
