+++ cmd/apc/workspace.go
@@ -9,7 +9,7 @@ import (
 	"github.com/astronomer/astro-cli/config"
 	"github.com/astronomer/astro-cli/houston"
 	"github.com/astronomer/astro-cli/pkg/logger"
-	"github.com/astronomer/astro-cli/software/workspace"
+	"github.com/astronomer/astro-cli/apc/workspace"
 )
 
 var (
